e2e/requirements-ray.txt
peft @ git+https://github.com/huggingface/peft@a18734d87aa9ae6b94b5bdde192b265bfad7c0b3
ray[serve]==2.9.3
torch==2.2.1
transformers @ git+https://github.com/huggingface/transformers@00c1d87a7d5c8dfb4554370983b5a3f7c069edd7
google-cloud-storage==2.14.0
ob-metaflow==2.11.4.9
scikit-learn==1.3.1
